backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
werkzeug>=3.1.3
pyjwt>=2.10.1
python-dotenv>=1.1.0
pytest>=8.3.5